#668: Fix SOTA chart display
